examples clear
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@1929 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ examples/After FX/clear.cmd
@@ -1,10 +1,8 @@
-pushd "%0\.."
+echo Clearing AfterFX examples in %CD%
 
-echo "Clearing AfterFX examples in %CD%"
+For /F "Tokens=*" %%I in ('dir /b "AE Project.aep*"') Do if not "%%I" == "AE Project.aep" del "%%I"
 
-del "AE Project.aep*.aep"
-rd /s /q "AE Project.aep Logs"
+if exist "AE Project.aep Logs" rd /s /q "AE Project.aep Logs"
+
 cd render
 del /f /q *
-
-popd
